+++ packages/plugin-packages/mashroom-portal-remote-messaging-app/src/js/components/SubscriptionPanel.tsx
@@ -7,37 +7,57 @@ import type {MashroomPortalMessageBus} from '@mashroom/mashroom-portal/type-defi
 import type {Subscription, ReceivedMessage} from '../types';
 
 type Props = {
-    subscription: Subscription,
+    privateUserTopicsSubscription: Subscription,
+    globalNotificationsSubscription: Subscription,
     messageBus: MashroomPortalMessageBus,
-    setSubscription: (subscription: Subscription) => void,
+    setPrivateUserTopicsSubscription: (subscription: Subscription) => void,
+    setGlobalNotificationsSubscription: (subscription: Subscription) => void,
     addReceivedMessage: (message: ReceivedMessage) => void,
 }
 
 export default class SubscriptionPanel extends PureComponent<Props> {
 
     componentDidMount(): void {
-        const { messageBus, setSubscription } = this.props;
+        const { messageBus, setPrivateUserTopicsSubscription, setGlobalNotificationsSubscription } = this.props;
         const userPrivateTopic = messageBus.getRemoteUserPrivateTopic();
+
+        const globalNotificationsTopic = `${messageBus.getRemotePrefix()}global-notifications`;
+        messageBus.subscribe(globalNotificationsTopic, this.onReceiveMessage.bind(this)).then(
+            () => {
+                setGlobalNotificationsSubscription({
+                    topic: globalNotificationsTopic,
+                    status: 'Success',
+                });
+            },
+            (error) => {
+                setPrivateUserTopicsSubscription({
+                    topic: globalNotificationsTopic,
+                    status: 'Error',
+                    errorMessage: JSON.stringify(error),
+                });
+            }
+        );
+
         if (!userPrivateTopic) {
             console.error('No messaging or websocket support, cannot subscribe to remote topic');
-            setSubscription({
+            setPrivateUserTopicsSubscription({
                 topic: '',
                 status: 'Error',
             });
             return;
         }
 
-        const topic = `${messageBus.getRemotePrefix()}${userPrivateTopic}/#`;
-        messageBus.subscribe(topic, this.onReceiveMessage.bind(this)).then(
+        const allUserPrivateTopics = `${messageBus.getRemotePrefix()}${userPrivateTopic}/#`;
+        messageBus.subscribe(allUserPrivateTopics, this.onReceiveMessage.bind(this)).then(
             () => {
-                setSubscription({
-                    topic,
+                setPrivateUserTopicsSubscription({
+                    topic: allUserPrivateTopics,
                     status: 'Success',
                 });
             },
             (error) => {
-                setSubscription({
-                    topic,
+                setPrivateUserTopicsSubscription({
+                    topic: allUserPrivateTopics,
                     status: 'Error',
                     errorMessage: JSON.stringify(error),
                 });
@@ -56,26 +76,37 @@ export default class SubscriptionPanel extends PureComponent<Props> {
     }
 
     render(): ReactNode {
-        const { subscription: { topic, status, errorMessage } } = this.props;
+        const { privateUserTopicsSubscription, globalNotificationsSubscription } = this.props;
 
         return (
             <div className='mashroom-remote-messaging-app-subscription-status'>
                 <div className='mashroom-remote-messaging-app-output-row'>
                     <div>
-                        <FormattedMessage id='subscribeTopic' />
+                        <FormattedMessage id='subscribeTopics' />
                     </div>
-                    <div className='remote-topic'>
-                        {topic}
-                    </div>
-                </div>
-                <div className='mashroom-remote-messaging-app-output-row'>
                     <div>
-                        <FormattedMessage id='subscriptionStatus' />
-                    </div>
-                    <div className={`${status === 'Error' ? 'mashroom-portal-ui-error-message' : ''}`}>
-                        {status}
-                        &nbsp;
-                        {errorMessage ? `(${errorMessage})` : ''}
+                        <ul>
+                            <li>
+                                <div className='remote-topic'>
+                                    {privateUserTopicsSubscription.topic}
+                                </div>
+                                <div className={`${privateUserTopicsSubscription.status === 'Error' ? 'mashroom-portal-ui-error-message' : ''}`}>
+                                    ({privateUserTopicsSubscription.status}
+                                    &nbsp;
+                                    {privateUserTopicsSubscription.errorMessage ? `(${privateUserTopicsSubscription.errorMessage})` : ''})
+                                </div>
+                            </li>
+                            <li>
+                                <div className='remote-topic'>
+                                    {globalNotificationsSubscription.topic}
+                                </div>
+                                <div className={`${globalNotificationsSubscription.status === 'Error' ? 'mashroom-portal-ui-error-message' : ''}`}>
+                                    ({globalNotificationsSubscription.status}
+                                    &nbsp;
+                                    {globalNotificationsSubscription.errorMessage ? `(${globalNotificationsSubscription.errorMessage})` : ''})
+                                </div>
+                            </li>
+                        </ul>
                     </div>
                 </div>
             </div>