+++ packages/plugin-packages/mashroom-portal-remote-messaging-app/src/js/store/actions.ts
@@ -6,14 +6,22 @@ import type {
     PublishedMessageStatus
 } from '../types';
 
-export const SET_SUBSCRIPTION = 'SET_SUBSCRIPTION';
+export const SET_PRIVATE_USER_TOPICS_SUBSCRIPTION = 'SET_PRIVATE_USER_TOPICS_SUBSCRIPTION';
+export const SET_GLOBAL_NOTIFICATIONS_SUBSCRIPTION = 'SET_GLOBAL_NOTIFICATIONS_SUBSCRIPTION';
 export const ADD_PUBLISHED_MESSAGE = 'ADD_PUBLISHED_MESSAGE';
 export const UPDATE_PUBLISHED_MESSAGE_STATUS = 'UPDATE_PUBLISHED_MESSAGE_STATUS';
 export const ADD_RECEIVED_MESSAGE = 'ADD_RECEIVED_MESSAGE';
 
-export const setSubscription = (subscription: Subscription) => {
+export const setPrivateUserTopicsSubscription = (subscription: Subscription) => {
     return {
-        type: SET_SUBSCRIPTION,
+        type: SET_PRIVATE_USER_TOPICS_SUBSCRIPTION,
+        subscription,
+    };
+};
+
+export const setGlobalNotificationsSubscription = (subscription: Subscription) => {
+    return {
+        type: SET_GLOBAL_NOTIFICATIONS_SUBSCRIPTION,
         subscription,
     };
 };